Correcion en migraciones

+++ database/migrations/2016_10_31_133759_create_menus_table.php
@@ -17,7 +17,7 @@ class CreateMenusTable extends Migration
             $table->increments('id');
             $table->integer('plato_id');
             $table->integer('tipo_ingreso_id');
-            $table->integer('cantidad');
+            $table->unsignedSmallInteger('cantidad');
             $table->date('fecha');
             $table->timestamps();
         });
